pyproject.toml
[project]
name = "rag-system"
version = "1.0.0"
description = "Local LLM-powered Retrieval-Augmented Generation system using Ollama and ChromaDB"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "pypdf>=3.17.0",
    "markdown>=3.5",
    "ollama>=0.4.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-playwright>=0.6.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
rag-system = "rag_system.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_system"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short -m 'not e2e'"
markers = ["e2e: end-to-end browser tests (require Playwright)"]
filterwarnings = [
    "ignore::DeprecationWarning:websockets",
]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
